frameworks/langgraph/requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
opentelemetry-api>=1.29.0
opentelemetry-sdk>=1.29.0
opentelemetry-exporter-otlp-proto-http>=1.29.0
opentelemetry-processor-baggage>=0.61b0
openinference-instrumentation-langchain
# test dependencies
pytest>=7.0.0
pytest-asyncio>=0.23.0
